+++ .gitlab-ci.yml
@@ -17,11 +17,11 @@ lambda-runtime:linuxX64Test:
     - trunk
     - external_pull_requests
 
-slackbot:linuxX64Test:
+webhook:linuxX64Test:
   image: gradle:7.4.2-jdk11
   stage: test
   needs: ["lambda-runtime:linuxX64Test", "lambda-runtime:jvmTest"]
-  script: gradle :slackbot:linuxX64Test
+  script: gradle :webhook:linuxX64Test
   only:
     - trunk
     - external_pull_requests
