+++ aws/rie-entry-script.sh
@@ -3,4 +3,4 @@ set -e
 set -u
 set -o pipefail
 
-exec /usr/local/bin/aws-lambda-rie /usr/local/bin/slackbot.kexe $@
+exec /usr/local/bin/aws-lambda-rie /usr/local/bin/webhook.kexe $@
